refactor: split internal/config/config.go by domain
Pure move: config.go (1494 lines, 45 declarations) split into
concern-named sibling files (config_strings_types, config_strings,
config_doors, config_login, config_theme, config_ftn, config_server,
config_v3net, config_events), following the same-package sibling-file
convention used by internal/menu/executor_*.go and
internal/configeditor/fields_*.go. Only import blocks changed; no
behavior, exported API, or test changes.
