+++ internal/config/config.go
@@ -1,1471 +1,11 @@
 package config
 
 import (
-	"encoding/json"
-	"fmt"
 	"log/slog"
-	"net"
 	"os"
-	"path/filepath"
-	"regexp"
-	"strconv"
 	"strings"
 	"time"
 )
-
-// StringsConfig holds all the configurable strings for the BBS.
-// This is the single source of truth.
-type StringsConfig struct {
-	ConnectionStr           string `json:"connectionStr"`
-	LockedBaudStr           string `json:"lockedBaudStr"`
-	ApplyAsNewStr           string `json:"applyAsNewStr"`
-	GetNupStr               string `json:"getNupStr"`
-	ChatRequestStr          string `json:"chatRequestStr"`
-	LeaveFBStr              string `json:"leaveFBStr"`
-	QuoteTitle              string `json:"quoteTitle"`
-	QuoteMessageStr         string `json:"quoteMessageStr"`
-	QuoteStartLine          string `json:"quoteStartLine"`
-	QuoteEndLine            string `json:"quoteEndLine"`
-	Erase5MsgsStr           string `json:"erase5MsgsStr"`
-	ChangeBoardStr          string `json:"changeBoardStr"`
-	NewscanBoardStr         string `json:"newscanBoardStr"`
-	PostOnBoardStr          string `json:"postOnBoardStr"`
-	MsgTitleStr             string `json:"msgTitleStr"`
-	MsgToStr                string `json:"msgToStr"`
-	MsgAnonStr              string `json:"msgAnonStr"`
-	AnonymousName           string `json:"anonymousName"`
-	SlashStr                string `json:"slashStr"`
-	NewScanningStr          string `json:"newScanningStr"`
-	ChangeFileAreaStr       string `json:"changeFileAreaStr"`
-	LogOffStr               string `json:"logOffStr"`
-	ChangeAutoMsgStr        string `json:"changeAutoMsgStr"`
-	NewUserNameStr          string `json:"newUserNameStr"`
-	CreateAPassword         string `json:"createAPassword"`
-	PauseString             string `json:"pauseString"`
-	WhatsYourAlias          string `json:"whatsYourAlias"`
-	WhatsYourPw             string `json:"whatsYourPw"`
-	SysopWorkingStr         string `json:"sysopWorkingStr"`
-	SysopInDos              string `json:"sysopInDos"`
-	SystemPasswordStr       string `json:"systemPasswordStr"`
-	DefPrompt               string `json:"defPrompt"`
-	EnterChat               string `json:"enterChat"`
-	ExitChat                string `json:"exitChat"`
-	SysOpIsIn               string `json:"sysOpIsIn"`
-	SysOpIsOut              string `json:"sysOpIsOut"`
-	HeaderStr               string `json:"headerStr"`
-	InfoformPrompt          string `json:"infoformPrompt"`
-	NewInfoFormPrompt       string `json:"newInfoFormPrompt"`
-	UserNotFound            string `json:"userNotFound"`
-	DesignNewPrompt         string `json:"designNewPrompt"`
-	YourCurrentPrompt       string `json:"yourCurrentPrompt"`
-	WantHotKeys             string `json:"wantHotKeys"`
-	WantRumors              string `json:"wantRumors"`
-	YourUserNum             string `json:"yourUserNum"`
-	WelcomeNewUser          string `json:"welcomeNewUser"`
-	EnterNumberHeader       string `json:"enterNumberHeader"`
-	EnterNumber             string `json:"enterNumber"`
-	EnterUserNote           string `json:"enterUserNote"`
-	CurFileArea             string `json:"curFileArea"`
-	EnterRealName           string `json:"enterRealName"`
-	ReEnterPassword         string `json:"reEnterPassword"`
-	QuoteTop                string `json:"quoteTop"`
-	QuoteBottom             string `json:"quoteBottom"`
-	AskOneLiner             string `json:"askOneLiner"`
-	OneLinerAnonymousPrompt string `json:"oneLinerAnonymousPrompt"`
-	EnterOneLiner           string `json:"enterOneLiner"`
-	OneLinerLegend          string `json:"One_Liner_Legend"`
-	NewScanDateStr          string `json:"newScanDateStr"`
-	AddBatchPrompt          string `json:"addBatchPrompt"`
-	ListUsers               string `json:"listUsers"`
-	ViewArchivePrompt       string `json:"viewArchivePrompt"`
-	AreaMsgNewScan          string `json:"areaMsgNewScan"`
-	GetInfoPrompt           string `json:"getInfoPrompt"`
-	MsgNewScanPrompt        string `json:"msgNewScanPrompt"`
-	TypeFilePrompt          string `json:"typeFilePrompt"`
-	ConfPrompt              string `json:"confPrompt"`
-	FileListPrompt          string `json:"fileListPrompt"`
-	UploadFileStr           string `json:"uploadFileStr"`
-	DownloadStr             string `json:"downloadStr"`
-	ListRange               string `json:"listRange"`
-	ContinueStr             string `json:"continueStr"`
-	ViewWhichForm           string `json:"viewWhichForm"`
-	CheckingUserBase        string `json:"checkingUserBase"`
-	NameAlreadyUsed         string `json:"nameAlreadyUsed"`
-	InvalidUserName         string `json:"invalidUserName"`
-	SysPwIs                 string `json:"sysPwIs"`
-	NotValidated            string `json:"notValidated"`
-	HaveMail                string `json:"haveMail"`
-	ReadMailNow             string `json:"readMailNow"`
-	DeleteNotice            string `json:"deleteNotice"`
-	HaveFeedback            string `json:"haveFeedback"`
-	ReadFeedback            string `json:"readFeedback"`
-	LoginNow                string `json:"loginNow"`
-	NewUsersWaiting         string `json:"newUsersWaiting"`
-	VoteOnNewUsers          string `json:"voteOnNewUsers"`
-	WrongPassword           string `json:"wrongPassword"`
-	YesPromptText           string `json:"yesPromptText"`
-	NoPromptText            string `json:"noPromptText"`
-	AbortMessagePrompt      string `json:"abortMessagePrompt"`
-
-	// Added from Page 5
-	AddBBSName          string `json:"addBBSName"`
-	AddBBSNumber        string `json:"addBBSNumber"`
-	AddBBSBaud          string `json:"addBBSBaud"`
-	AddBBSSoftware      string `json:"addBBSSoftware"`
-	AddExtendedBBSDescr string `json:"addExtendedBBSDescr"`
-	BBSEntryAdded       string `json:"bbsEntryAdded"`
-	ViewNextDescrip     string `json:"viewNextDescrip"`
-	JoinedMsgConf       string `json:"joinedMsgConf"`
-	JoinedFileConf      string `json:"joinedFileConf"`
-	WhosBeingVotedOn    string `json:"whosBeingVotedOn"`
-	NumYesVotes         string `json:"numYesVotes"`
-	NumNoVotes          string `json:"numNoVotes"`
-	NUVCommentHeader    string `json:"nuvCommentHeader"`
-
-	// Added from Page 6
-	EnterNUVCommentPrompt string `json:"enterNUVCommentPrompt"`
-	NUVVotePrompt         string `json:"nuvVotePrompt"`
-	YesVoteCast           string `json:"yesVoteCast"`
-	NoVoteCast            string `json:"noVoteCast"`
-	NoNewUsersPending     string `json:"noNewUsersPending"`
-	EnterRumorTitle       string `json:"enterRumorTitle"`
-	AddRumorAnonymous     string `json:"addRumorAnonymous"`
-	EnterRumorLevel       string `json:"enterRumorLevel"`
-	EnterRumorPrompt      string `json:"enterRumorPrompt"`
-	RumorAdded            string `json:"rumorAdded"`
-	ListRumorsPrompt      string `json:"listRumorsPrompt"`
-	SendMailToWho         string `json:"sendMailToWho"`
-	CarbonCopyMail        string `json:"carbonCopyMail"`
-	NotifyEMail           string `json:"notifyEMail"`
-	EMailAnnouncement     string `json:"eMailAnnouncement"`
-	SysOpNotHere          string `json:"sysOpNotHere"`
-	ChatCostsHeader       string `json:"chatCostsHeader"`
-	StillWantToTry        string `json:"stillWantToTry"`
-	NotEnoughFPPoints     string `json:"notEnoughFPPoints"`
-	ChatCallOff           string `json:"chatCallOff"`
-
-	// Added from Page 7
-	ChatCallOn          string `json:"chatCallOn"`
-	FeedbackSent        string `json:"feedbackSent"`
-	YouHaveReadMail     string `json:"youHaveReadMail"`
-	DeleteMailNow       string `json:"deleteMailNow"`
-	CurrentMailNone     string `json:"currentMailNone"`
-	CurrentMailWaiting  string `json:"currentMailWaiting"`
-	PickMailHeader      string `json:"pickMailHeader"`
-	ListTitleType       string `json:"listTitleType"`
-	NoMoreTitles        string `json:"noMoreTitles"`
-	ListTitlesToYou     string `json:"listTitlesToYou"`
-	SubDoesNotExist     string `json:"subDoesNotExist"`
-	MsgNewScanAborted   string `json:"msgNewScanAborted"`
-	MsgReadingPrompt    string `json:"msgReadingPrompt"`
-	CurrentSubNewScan   string `json:"currentSubNewScan"`
-	JumpToMessageNum    string `json:"jumpToMessageNum"`
-	PostingQWKMsg       string `json:"postingQWKMsg"`
-	TotalQWKAdded       string `json:"totalQWKAdded"`
-	SendQWKPacketPrompt string `json:"sendQWKPacketPrompt"`
-
-	// Added from Page 8
-	ThreadWhichWay       string `json:"threadWhichWay"`
-	AutoValidatingFile   string `json:"autoValidatingFile"`
-	FileIsWorth          string `json:"fileIsWorth"`
-	GrantingUserFP       string `json:"grantingUserFP"`
-	FileIsOffline        string `json:"fileIsOffline"`
-	CrashedFile          string `json:"crashedFile"`
-	BadBaudRate          string `json:"badBaudRate"`
-	UnvalidatedFile      string `json:"unvalidatedFile"`
-	SpecialFile          string `json:"specialFile"`
-	NoDownloadsHere      string `json:"noDownloadsHere"`
-	PrivateFile          string `json:"privateFile"`
-	FilePassword         string `json:"filePassword"`
-	WrongFilePW          string `json:"wrongFilePW"`
-	FileNewScanPrompt    string `json:"fileNewScanPrompt"`
-	InvalidArea          string `json:"invalidArea"`
-	UntaggingBatchFile   string `json:"untaggingBatchFile"`
-	FileExtractionPrompt string `json:"fileExtractionPrompt"`
-
-	// Added from Page 9
-	BadUDRatio             string `json:"badUDRatio"`
-	BadUDKRatio            string `json:"badUDKRatio"`
-	ExceededDailyKBLimit   string `json:"exceededDailyKBLimit"`
-	FilePointCommision     string `json:"filePointCommision"`
-	SuccessfulDownload     string `json:"successfulDownload"`
-	FileCrashSave          string `json:"fileCrashSave"`
-	InvalidFilename        string `json:"invalidFilename"`
-	AlreadyEnteredFilename string `json:"alreadyEnteredFilename"`
-	FileAlreadyExists      string `json:"fileAlreadyExists"`
-	EnterFileDescription   string `json:"enterFileDescription"`
-	ExtendedUploadSetup    string `json:"extendedUploadSetup"`
-	ReEnterFileDescrip     string `json:"reEnterFileDescrip"`
-	NotifyIfDownloaded     string `json:"notifyIfDownloaded"`
-	FiftyFilesMaximum      string `json:"fiftyFilesMaximum"`
-	YouCantDownloadHere    string `json:"youCantDownloadHere"`
-	FileAlreadyMarked      string `json:"fileAlreadyMarked"`
-	NotEnoughFP            string `json:"notEnoughFP"`
-	FileAreaPassword       string `json:"fileAreaPassword"`
-
-	// Download/batch workflow strings (V3-specific)
-	BatchQueueEmpty        string `json:"batchQueueEmpty"`
-	BatchClearedFormat     string `json:"batchClearedFormat"`
-	AddedToBatchFormat     string `json:"addedToBatchFormat"`
-	NoFilesTagged          string `json:"noFilesTagged"`
-	BatchCountFormat       string `json:"batchCountFormat"`
-	FilesResolveError      string `json:"filesResolveError"`
-	DownloadFinishedFormat string `json:"downloadFinishedFormat"`
-	FileAreaNotFound       string `json:"fileAreaNotFound"`
-	SaveUserError          string `json:"saveUserError"`
-
-	QuotePrefix string `json:"QuotePrefix"`
-
-	// Chat strings (V3-specific)
-	ChatHeader        string `json:"chatHeader"`
-	ChatSeparator     string `json:"chatSeparator"`
-	ChatUserEntered   string `json:"chatUserEntered"`
-	ChatUserLeft      string `json:"chatUserLeft"`
-	ChatSystemPrefix  string `json:"chatSystemPrefix"`
-	ChatMessageFormat string `json:"chatMessageFormat"`
-
-	ChatNetworkPickerHeader string `json:"chatNetworkPickerHeader"`
-	ChatNetworkPickerEntry  string `json:"chatNetworkPickerEntry"`
-	ChatRoomListHeader      string `json:"chatRoomListHeader"`
-	ChatRoomListEntry       string `json:"chatRoomListEntry"`
-	ChatPrivateMsgFormat    string `json:"chatPrivateMsgFormat"`
-	ChatJoinMsg             string `json:"chatJoinMsg"`
-	ChatLeaveMsg            string `json:"chatLeaveMsg"`
-	ChatTopicMsg            string `json:"chatTopicMsg"`
-	ChatReconnected         string `json:"chatReconnected"`
-
-	// Page strings (V3-specific)
-	PageOnlineNodesHeader string `json:"pageOnlineNodesHeader"`
-	PageNodeListEntry     string `json:"pageNodeListEntry"`
-	PageWhichNodePrompt   string `json:"pageWhichNodePrompt"`
-	PageMessagePrompt     string `json:"pageMessagePrompt"`
-	PageMessageFormat     string `json:"pageMessageFormat"`
-	PageSent              string `json:"pageSent"`
-	PageCancelled         string `json:"pageCancelled"`
-	PageInvalidNode       string `json:"pageInvalidNode"`
-	PageSelfError         string `json:"pageSelfError"`
-	PageNodeOffline       string `json:"pageNodeOffline"`
-
-	// Newuser strings (V3-specific)
-	NewUsersClosedStr       string `json:"newUsersClosedStr"`
-	NewUserLocationPrompt   string `json:"newUserLocationPrompt"`
-	NewUserPasswordTooShort string `json:"newUserPasswordTooShort"`
-	NewUserPasswordMismatch string `json:"newUserPasswordMismatch"`
-	NewUserInvalidRealName  string `json:"newUserInvalidRealName"`
-	NewUserTooManyAttempts  string `json:"newUserTooManyAttempts"`
-	NewUserAccountCreated   string `json:"newUserAccountCreated"`
-	NewUserCreationError    string `json:"newUserCreationError"`
-	NewUserMaybeAnotherTime string `json:"newUserMaybeAnotherTime"`
-
-	// System stats strings (V3-specific)
-	StatsBBSName     string `json:"statsBBSName"`
-	StatsSysOp       string `json:"statsSysOp"`
-	StatsVersion     string `json:"statsVersion"`
-	StatsTotalUsers  string `json:"statsTotalUsers"`
-	StatsTotalCalls  string `json:"statsTotalCalls"`
-	StatsTotalMsgs   string `json:"statsTotalMessages"`
-	StatsTotalFiles  string `json:"statsTotalFiles"`
-	StatsActiveNodes string `json:"statsActiveNodes"`
-	StatsDate        string `json:"statsDate"`
-	StatsTime        string `json:"statsTime"`
-
-	// User config strings (V3-specific)
-	CfgToggleOn            string `json:"cfgToggleOn"`
-	CfgToggleOff           string `json:"cfgToggleOff"`
-	CfgToggleFormat        string `json:"cfgToggleFormat"`
-	CfgSaveError           string `json:"cfgSaveError"`
-	CfgScreenWidthPrompt   string `json:"cfgScreenWidthPrompt"`
-	CfgScreenWidthInvalid  string `json:"cfgScreenWidthInvalid"`
-	CfgScreenWidthSet      string `json:"cfgScreenWidthSet"`
-	CfgScreenHeightPrompt  string `json:"cfgScreenHeightPrompt"`
-	CfgScreenHeightInvalid string `json:"cfgScreenHeightInvalid"`
-	CfgScreenHeightSet     string `json:"cfgScreenHeightSet"`
-	CfgTermTypeSet         string `json:"cfgTermTypeSet"`
-	CfgStringPrompt        string `json:"cfgStringPrompt"`
-	CfgStringPromptCurrent string `json:"cfgStringPromptCurrent"`
-	CfgStringUpdated       string `json:"cfgStringUpdated"`
-	CfgCurrentPwPrompt     string `json:"cfgCurrentPwPrompt"`
-	CfgIncorrectPw         string `json:"cfgIncorrectPw"`
-	CfgPasswordChanged     string `json:"cfgPasswordChanged"`
-	CfgColorSelectPrompt   string `json:"cfgColorSelectPrompt"`
-	CfgColorInputPrompt    string `json:"cfgColorInputPrompt"`
-	CfgColorInvalid        string `json:"cfgColorInvalid"`
-	CfgColorSet            string `json:"cfgColorSet"`
-	CfgCustomPromptHelp    string `json:"cfgCustomPromptHelp"`
-	CfgViewScreenWidth     string `json:"cfgViewScreenWidth"`
-	CfgViewScreenHeight    string `json:"cfgViewScreenHeight"`
-	CfgViewTermType        string `json:"cfgViewTermType"`
-	CfgViewHotKeys         string `json:"cfgViewHotKeys"`
-	CfgViewMorePrompts     string `json:"cfgViewMorePrompts"`
-	CfgViewMsgHeader       string `json:"cfgViewMsgHeader"`
-	CfgViewCustomPrompt    string `json:"cfgViewCustomPrompt"`
-	CfgViewPromptColor     string `json:"cfgViewPromptColor"`
-	CfgViewTextColor       string `json:"cfgViewTextColor"`
-	CfgViewText2Color      string `json:"cfgViewText2Color"`
-	CfgViewBarColor        string `json:"cfgViewBarColor"`
-	CfgViewRealName        string `json:"cfgViewRealName"`
-	CfgViewNote            string `json:"cfgViewNote"`
-	CfgViewFileListMode    string `json:"cfgViewFileListMode"`
-	CfgFileListModeSet     string `json:"cfgFileListModeSet"`
-
-	// Message reader strings (V3-specific)
-	MsgEndOfMessages     string `json:"msgEndOfMessages"`
-	MsgFirstMessage      string `json:"msgFirstMessage"`
-	MsgMailReplyDeferred string `json:"msgMailReplyDeferred"`
-	MsgListDeferred      string `json:"msgListDeferred"`
-	MsgHdrLoadError      string `json:"msgHdrLoadError"`
-	MsgBoardInfoFormat   string `json:"msgBoardInfoFormat"`
-	MsgScrollPercent     string `json:"msgScrollPercent"`
-	MsgNewScanSuffix     string `json:"msgNewScanSuffix"`
-	MsgReadingSuffix     string `json:"msgReadingSuffix"`
-	MsgThreadPrompt      string `json:"msgThreadPrompt"`
-	MsgNoThreadFound     string `json:"msgNoThreadFound"`
-	MsgJumpPrompt        string `json:"msgJumpPrompt"`
-	MsgInvalidMsgNum     string `json:"msgInvalidMsgNum"`
-	MsgReplySubjectEmpty string `json:"msgReplySubjectEmpty"`
-	MsgLaunchingEditor   string `json:"msgLaunchingEditor"`
-	MsgReplyCancelled    string `json:"msgReplyCancelled"`
-	MsgReplyError        string `json:"msgReplyError"`
-	MsgReplySuccess      string `json:"msgReplySuccess"`
-	MsgEditorError       string `json:"msgEditorError"`
-
-	// Message scan strings (V3-specific)
-	ScanDateLine            string `json:"scanDateLine"`
-	ScanToLine              string `json:"scanToLine"`
-	ScanFromLine            string `json:"scanFromLine"`
-	ScanRangeLine           string `json:"scanRangeLine"`
-	ScanUpdateLine          string `json:"scanUpdateLine"`
-	ScanWhichLine           string `json:"scanWhichLine"`
-	ScanAbortLine           string `json:"scanAbortLine"`
-	ScanSelectionPrompt     string `json:"scanSelectionPrompt"`
-	ScanDatePrompt          string `json:"scanDatePrompt"`
-	ScanToPrompt            string `json:"scanToPrompt"`
-	ScanFromPrompt          string `json:"scanFromPrompt"`
-	ScanRangeStartPrompt    string `json:"scanRangeStartPrompt"`
-	ScanRangeEndPrompt      string `json:"scanRangeEndPrompt"`
-	ScanWhichPrompt         string `json:"scanWhichPrompt"`
-	ScanHeader              string `json:"scanHeader"`
-	ScanNoAreaSelected      string `json:"scanNoAreaSelected"`
-	ScanNoMessages          string `json:"scanNoMessages"`
-	ScanNoTaggedAreas       string `json:"scanNoTaggedAreas"`
-	NewscanNewNetworkPrompt string `json:"newscanNewNetworkPrompt"`
-	ScanAreaProgress        string `json:"scanAreaProgress"`
-	ScanComplete            string `json:"scanComplete"`
-	ScanLoginRequired       string `json:"scanLoginRequired"`
-	ScanNoAreasAvailable    string `json:"scanNoAreasAvailable"`
-	ScanNoAccessibleAreas   string `json:"scanNoAccessibleAreas"`
-	ScanConfigSaved         string `json:"scanConfigSaved"`
-	ScanConfigError         string `json:"scanConfigError"`
-	ScanConfigLoginRequired string `json:"scanConfigLoginRequired"`
-
-	// Update newscan pointers strings
-	UpdatePtrsLoginRequired string `json:"updatePtrsLoginRequired"`
-	UpdatePtrsCancelled     string `json:"updatePtrsCancelled"`
-	UpdatePtrsDatePrompt    string `json:"updatePtrsDatePrompt"`
-	UpdatePtrsScopePrompt   string `json:"updatePtrsScopePrompt"`
-	UpdatePtrsSuccess       string `json:"updatePtrsSuccess"`
-	UpdatePtrsError         string `json:"updatePtrsError"`
-
-	// Message list strings (V3-specific)
-	MsgListLoginRequired  string `json:"msgListLoginRequired"`
-	MsgListNoAreaSelected string `json:"msgListNoAreaSelected"`
-	MsgListAreaNotFound   string `json:"msgListAreaNotFound"`
-	MsgListLoadError      string `json:"msgListLoadError"`
-	MsgListNoMessages     string `json:"msgListNoMessages"`
-
-	// File viewer strings (V3-specific)
-	FileNoAreaSelected string `json:"fileNoAreaSelected"`
-	FilePromptFormat   string `json:"filePromptFormat"`
-	FileNotFoundFormat string `json:"fileNotFoundFormat"`
-	FileLocateError    string `json:"fileLocateError"`
-	FileViewingHeader  string `json:"fileViewingHeader"`
-	FileEndOfFile      string `json:"fileEndOfFile"`
-	FileMorePrompt     string `json:"fileMorePrompt"`
-	FilePausePrompt    string `json:"filePausePrompt"`
-	FileOpenError      string `json:"fileOpenError"`
-
-	// File search strings
-	SearchFilesPrompt    string `json:"searchFilesPrompt"`
-	SearchFilesMinChars  string `json:"searchFilesMinChars"`
-	SearchNoResults      string `json:"searchNoResults"`
-	SearchResultsHeader  string `json:"searchResultsHeader"`
-	SearchResultFormat   string `json:"searchResultFormat"`
-	SearchResultsSummary string `json:"searchResultsSummary"`
-
-	// File info strings
-	FileInfoPrompt string `json:"fileInfoPrompt"`
-	FileInfoHeader string `json:"fileInfoHeader"`
-
-	// File newscan strings
-	FileNewscanHeader   string `json:"fileNewscanHeader"`
-	FileNewscanAreaHdr  string `json:"fileNewscanAreaHdr"`
-	FileNewscanNoNew    string `json:"fileNewscanNoNew"`
-	FileNewscanComplete string `json:"fileNewscanComplete"`
-
-	// Sysop file review strings
-	SysopReviewHeader  string `json:"sysopReviewHeader"`
-	SysopReviewPrompt  string `json:"sysopReviewPrompt"`
-	SysopReviewMarked  string `json:"sysopReviewMarked"`
-	SysopReviewNoFiles string `json:"sysopReviewNoFiles"`
-	SysopReviewScanAll string `json:"sysopReviewScanAll"`
-	SysopReviewRenamed string `json:"sysopReviewRenamed"`
-
-	// File newscan config strings
-	FileNewscanConfigHeader string `json:"fileNewscanConfigHeader"`
-	FileNewscanConfigPrompt string `json:"fileNewscanConfigPrompt"`
-	FileNewscanConfigSaved  string `json:"fileNewscanConfigSaved"`
-
-	// Column config strings
-	CfgFileColumnsHeader string `json:"cfgFileColumnsHeader"`
-	CfgFileColumnsToggle string `json:"cfgFileColumnsToggle"`
-	CfgFileColumnsSaved  string `json:"cfgFileColumnsSaved"`
-
-	// Want list strings
-	WantListPrompt       string `json:"wantListPrompt"`
-	WantListReasonPrompt string `json:"wantListReasonPrompt"`
-	WantListSubmitted    string `json:"wantListSubmitted"`
-	WantListEmpty        string `json:"wantListEmpty"`
-	WantListHeader       string `json:"wantListHeader"`
-	WantListCleared      string `json:"wantListCleared"`
-
-	// Door handler strings (V3-specific)
-	DoorDropfileError     string `json:"doorDropfileError"`
-	DoorErrorFormat       string `json:"doorErrorFormat"`
-	DoorLoginRequired     string `json:"doorLoginRequired"`
-	DoorPrompt            string `json:"doorPrompt"`
-	DoorNotFoundFormat    string `json:"doorNotFoundFormat"`
-	DoorNoneConfigured    string `json:"doorNoneConfigured"`
-	DoorTemplateError     string `json:"doorTemplateError"`
-	DoorInfoLoginRequired string `json:"doorInfoLoginRequired"`
-	DoorAccessDenied      string `json:"doorAccessDenied"`
-	DoorBusyFormat        string `json:"doorBusyFormat"`
-
-	// Matrix strings (V3-specific)
-	MatrixDisconnecting       string `json:"matrixDisconnecting"`
-	MatrixCheckAccessPrompt   string `json:"matrixCheckAccessPrompt"`
-	MatrixUserNotFound        string `json:"matrixUserNotFound"`
-	MatrixAccountValidated    string `json:"matrixAccountValidated"`
-	MatrixAccountNotValidated string `json:"matrixAccountNotValidated"`
-	IdleTimeout               string `json:"idleTimeout"`
-
-	// Conference menu strings (V3-specific)
-	ConfLoginRequired           string `json:"confLoginRequired"`
-	ConfNoConferences           string `json:"confNoConferences"`
-	ConfNotFound                string `json:"confNotFound"`
-	ConfNoAccessibleAreas       string `json:"confNoAccessibleAreas"`
-	ConfCurrentAreaFormat       string `json:"confCurrentAreaFormat"`
-	ConfTemplateError           string `json:"confTemplateError"`
-	ConfNavLoginRequired        string `json:"confNavLoginRequired"`
-	ConfNoAccessibleConferences string `json:"confNoAccessibleConferences"`
-	ConfNoAccessibleMsgAreas    string `json:"confNoAccessibleMsgAreas"`
-	ConfAreaTemplateError       string `json:"confAreaTemplateError"`
-	ConfCurrentConfFormat       string `json:"confCurrentConfFormat"`
-	ConfNoAccessibleConfs       string `json:"confNoAccessibleConfs"`
-
-	// Executor strings (V3-specific)
-	ExecUnknownCommand      string `json:"execUnknownCommand"`
-	ExecReadmailLogin       string `json:"execReadmailLogin"`
-	ExecReadmailPlaceholder string `json:"execReadmailPlaceholder"`
-	ExecDoorLogin           string `json:"execDoorLogin"`
-	ExecDoorNotConfigured   string `json:"execDoorNotConfigured"`
-	ExecGoodbye             string `json:"execGoodbye"`
-	ExecStatsLogin          string `json:"execStatsLogin"`
-	ExecStatsError          string `json:"execStatsError"`
-	ExecOnelinerTemplateErr string `json:"execOnelinerTemplateErr"`
-	ExecOnelinerColorError  string `json:"execOnelinerColorError"`
-	ExecOnelinerWriteError  string `json:"execOnelinerWriteError"`
-	ExecOnelinerAdded       string `json:"execOnelinerAdded"`
-	ExecOnelinerEmpty       string `json:"execOnelinerEmpty"`
-	ExecAlreadyLoggedIn     string `json:"execAlreadyLoggedIn"`
-	ExecUsernamePrompt      string `json:"execUsernamePrompt"`
-	ExecPasswordPrompt      string `json:"execPasswordPrompt"`
-	ExecAbortLoginPrompt    string `json:"execAbortLoginPrompt"`
-	ExecLoginCancelled      string `json:"execLoginCancelled"`
-	ExecIPLockout           string `json:"execIPLockout"`
-	ExecLoginIncorrect      string `json:"execLoginIncorrect"`
-	ExecNotValidated        string `json:"execNotValidated"`
-	ExecMenuLoadError       string `json:"execMenuLoadError"`
-	ExecMenuPasswordPrompt  string `json:"execMenuPasswordPrompt"`
-	ExecPasswordAccepted    string `json:"execPasswordAccepted"`
-	ExecIncorrectPassword   string `json:"execIncorrectPassword"`
-	ExecTooManyAttempts     string `json:"execTooManyAttempts"`
-	ExecAccessDenied        string `json:"execAccessDenied"`
-	ExecLastcallTemplateErr string `json:"execLastcallTemplateErr"`
-	ExecFileLoadError       string `json:"execFileLoadError"`
-	ExecNoNewMail           string `json:"execNoNewMail"`
-	ExecNewMailCount        string `json:"execNewMailCount"`
-	ExecUserlistTemplateErr string `json:"execUserlistTemplateErr"`
-	ExecPendingValidation   string `json:"execPendingValidation"`
-	ExecRunCommandError     string `json:"execRunCommandError"`
-	ExecRunCommandNotFound  string `json:"execRunCommandNotFound"`
-	ExecRunDoorError        string `json:"execRunDoorError"`
-	ExecLoginCriticalError  string `json:"execLoginCriticalError"`
-	ExecVersionString       string `json:"execVersionString"`
-
-	// Terminal size mismatch prompts (shown post-login when detected size differs from saved)
-	TermSizeNewDetectedPrompt    string `json:"termSizeNewDetectedPrompt"`
-	TermSizeUpdateDefaultsPrompt string `json:"termSizeUpdateDefaultsPrompt"`
-
-	// Invisible login prompt (shown to SysOp/CoSysOp after authentication)
-	InvisibleLogonPrompt string `json:"invisibleLogonPrompt"`
-
-	// Default Colors (|C1 - |C7 map to these)
-	DefColor1 uint8 `json:"defColor1"`
-	DefColor2 uint8 `json:"defColor2"`
-	DefColor3 uint8 `json:"defColor3"`
-	DefColor4 uint8 `json:"defColor4"`
-	DefColor5 uint8 `json:"defColor5"`
-	DefColor6 uint8 `json:"defColor6"`
-	DefColor7 uint8 `json:"defColor7"`
-}
-
-// GlobalStrings holds the loaded configuration.
-// var GlobalStrings StringsConfig // Removed duplicate
-
-// LoadStrings loads the string configuration from a JSON file.
-func LoadStrings(configPath string) (StringsConfig, error) { // Return the loaded config directly
-	filePath := filepath.Join(configPath, "strings.json")
-	slog.Info("loading strings configuration", "path", filePath)
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		slog.Error("failed to read strings file", "path", filePath, "error", err)
-		return StringsConfig{}, fmt.Errorf("failed to read strings file %s: %w", filePath, err)
-	}
-
-	var loadedConfig StringsConfig // Load into a local variable
-	err = json.Unmarshal(data, &loadedConfig)
-	if err != nil {
-		slog.Error("failed to parse strings JSON", "path", filePath, "error", err)
-		return StringsConfig{}, fmt.Errorf("failed to parse strings JSON from %s: %w", filePath, err)
-	}
-
-	applyStringDefaults(&loadedConfig)
-	slog.Info("successfully loaded strings configuration")
-	return loadedConfig, nil // Return the loaded struct
-}
-
-// applyStringDefaults fills in default values for any string keys that are empty.
-func applyStringDefaults(c *StringsConfig) {
-	d := func(field *string, val string) {
-		if *field == "" {
-			*field = val
-		}
-	}
-
-	// File search
-	d(&c.SearchFilesPrompt, "\r\n|15Enter search text |07(min 3 chars)|15: |07")
-	d(&c.SearchFilesMinChars, "\r\n|12Search text must be at least 3 characters.|07\r\n")
-	d(&c.SearchNoResults, "\r\n|14No files found matching your search.|07\r\n")
-	d(&c.SearchResultsHeader, "\r\n|15Search results for: |11%s|07\r\n|08────────────────────────────────────────────────────────────────|07\r\n")
-	d(&c.SearchResultsSummary, "\r\n|15%d file(s) found.|07\r\n")
-
-	// File info
-	d(&c.FileInfoPrompt, "\r\n|15Filename to view info|15: |07")
-	d(&c.FileInfoHeader, "\r\n|08════════════════════════════════════════|07\r\n|15File Information|07\r\n|08════════════════════════════════════════|07")
-
-	// File newscan
-	d(&c.FileNewscanHeader, "\r\n|15File Newscan|07\r\n|08────────────────────────────────────────|07\r\n")
-	d(&c.FileNewscanAreaHdr, "\r\n|11%s |07(%d new files)|07\r\n")
-	d(&c.FileNewscanNoNew, "\r\n|14No new files found since your last login.|07\r\n")
-	d(&c.FileNewscanComplete, "\r\n|15Newscan complete. |11%d|15 new file(s) found.|07\r\n")
-
-	// File newscan config
-	d(&c.FileNewscanConfigHeader, "|15File Newscan Configuration|07\r\n|08────────────────────────────────────────|07\r\n|07Tag areas to scan for new files|07\r\n")
-	d(&c.FileNewscanConfigSaved, "\r\n|15File newscan config saved. |11%d|15 area(s) tagged.|07\r\n")
-
-	// Sysop file review
-	d(&c.SysopReviewHeader, "|15SysOp File Review|07")
-	d(&c.SysopReviewPrompt, "\r\n|07[|15C|07]hange Desc  [|15R|07]ename  [|15D|07]elete  [|15M|07]ove  [|15S|07]kip  [|15Q|07]uit: |15")
-	d(&c.SysopReviewMarked, "|10File marked as reviewed.|07")
-	d(&c.SysopReviewNoFiles, "\r\n|14No unreviewed files.|07\r\n")
-	d(&c.SysopReviewScanAll, "\r\n|07Scan all areas? [|15Y|07/|15N|07]: |15")
-	d(&c.SysopReviewRenamed, "|10File renamed successfully.|07")
-
-	// Want list
-	d(&c.WantListPrompt, "\r\n|15Filename to request|15: |07")
-	d(&c.WantListReasonPrompt, "\r\n|07Reason |08(optional)|07: |07")
-	d(&c.WantListSubmitted, "\r\n|10Request submitted.|07\r\n")
-	d(&c.WantListEmpty, "\r\n|14No file requests.|07\r\n")
-	d(&c.WantListHeader, "\r\n|15File Want List|07\r\n|08────────────────────────────────────────|07\r\n")
-	d(&c.WantListCleared, "\r\n|10Want list cleared.|07\r\n")
-
-	// Column config
-	d(&c.CfgFileColumnsHeader, "\r\n|15File Listing Columns|07\r\n|08────────────────────────────────────────|07\r\n")
-	d(&c.CfgFileColumnsToggle, "  |15[%s]|07 %-12s : %s\r\n")
-	d(&c.CfgFileColumnsSaved, "\r\n|10Column preferences saved.|07\r\n")
-
-	// Door access control
-	d(&c.DoorAccessDenied, "\r\n|14Access denied to door: |11%s|07\r\n")
-	d(&c.DoorBusyFormat, "\r\n|14Door is currently in use: |11%s|07\r\n")
-}
-
-// DoorConfig defines the configuration for a single external door program.
-type DoorConfig struct {
-	Code                string            `json:"code"`                            // Unique internal code used in DOOR:CODE commands (uppercase slug)
-	Name                string            `json:"name"`                            // Display label shown to users (free-form, case preserved)
-	WorkingDirectory    string            `json:"working_directory,omitempty"`     // Directory to run the command in (optional)
-	Commands            []string          `json:"commands,omitempty"`              // Commands to execute (native: [0]=executable, [1:]=args; DOS: batch lines)
-	DropfileType        string            `json:"dropfile_type,omitempty"`         // Type of dropfile ("DOOR.SYS", "CHAIN.TXT", "NONE") (optional, defaults to NONE)
-	DropfileLocation    string            `json:"dropfile_location,omitempty"`     // Where to write dropfile: "startup" (working dir, default) or "node" (per-node temp dir)
-	DropfileCase        string            `json:"dropfile_case,omitempty"`         // Dropfile filename case: "upper" (default) or "lower"
-	IOMode              string            `json:"io_mode,omitempty"`               // I/O handling ("STDIO", "SOCKET") (optional, defaults to STDIO)
-	RequiresRawTerminal bool              `json:"requires_raw_terminal,omitempty"` // Whether the BBS should attempt to put the terminal in raw mode (optional, defaults to false)
-	UseShell            bool              `json:"use_shell,omitempty"`             // Wrap command in /bin/sh -c (Linux) or cmd /c (Windows)
-	SingleInstance      bool              `json:"single_instance,omitempty"`       // Only allow one node to run this door at a time
-	MinAccessLevel      int               `json:"min_access_level,omitempty"`      // Minimum user access level required (0 = no restriction)
-	CleanupCommand      string            `json:"cleanup_command,omitempty"`       // Command to run after door exits (optional)
-	CleanupArgs         []string          `json:"cleanup_args,omitempty"`          // Arguments for cleanup command (supports placeholders)
-	EnvironmentVars     map[string]string `json:"environment_variables,omitempty"` // Additional environment variables (optional)
-	// Script door fields
-	Type         string   `json:"type,omitempty"`          // "synchronet_js", "v3_script", or empty (legacy native/DOS)
-	Script       string   `json:"script,omitempty"`        // Main JS file to execute (relative to working_directory)
-	LibraryPaths []string `json:"library_paths,omitempty"` // Search paths for load()/require()
-	Args         []string `json:"args,omitempty"`          // Script arguments (available as argv in JS)
-	ExecDir      string   `json:"exec_dir,omitempty"`      // Synchronet exec directory (system.exec_dir)
-	// DOS door fields
-	IsDOS        bool   `json:"is_dos,omitempty"`        // true = DOS door launched via a DOS emulator
-	DriveCPath   string `json:"drive_c_path,omitempty"`  // Path to drive_c directory (default: ~/.dosemu/drive_c)
-	DOSEmulator  string `json:"dos_emulator,omitempty"`  // Emulator to use: "auto" (default) or "dosemu"
-	FossilDriver string `json:"fossil_driver,omitempty"` // DOS FOSSIL driver command (e.g. "C:\\UTILS\\X00.EXE eliminate")
-	// dosemu2-specific fields (Linux x86 only)
-	DosemuConfig string `json:"dosemu_config,omitempty"` // Path to custom .dosemurc (optional)
-}
-
-// doorCodeRE validates a door code after uppercasing: the code keys the door
-// registry and appears in DOOR:CODE menu commands, so it must be a short slug
-// with no spaces or punctuation.
-var doorCodeRE = regexp.MustCompile(`^[A-Z0-9_-]{1,16}$`)
-
-// NormalizeDoorCode trims and uppercases a door code and validates it against
-// the required format. Both the loader and the config editor use this, so the
-// contract is enforced identically for hand-edited and TUI-edited configs.
-func NormalizeDoorCode(code string) (string, error) {
-	code = strings.ToUpper(strings.TrimSpace(code))
-	if !doorCodeRE.MatchString(code) {
-		return "", fmt.Errorf("door code must be 1-16 chars: A-Z, 0-9, _ or -")
-	}
-	return code, nil
-}
-
-// LoadDoors loads the door configuration from the specified file path.
-func LoadDoors(filePath string) (map[string]DoorConfig, error) {
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		// If the file doesn't exist, return an empty map and no error, as doors are optional.
-		if os.IsNotExist(err) {
-			return make(map[string]DoorConfig), nil
-		}
-		return nil, fmt.Errorf("failed to read doors file %s: %w", filePath, err)
-	}
-
-	var doors []DoorConfig
-	err = json.Unmarshal(data, &doors)
-	if err != nil {
-		return nil, fmt.Errorf("failed to unmarshal doors JSON from %s: %w", filePath, err)
-	}
-
-	// Key by uppercased Code: menu lookups uppercase the door code before
-	// consulting the registry, so codes must normalize or the door is
-	// unreachable. Name is a display label and is left untouched.
-	doorMap := make(map[string]DoorConfig)
-	for _, door := range doors {
-		code, err := NormalizeDoorCode(door.Code)
-		if err != nil {
-			return nil, fmt.Errorf("door %q in %s: %w", door.Name, filePath, err)
-		}
-		if _, exists := doorMap[code]; exists {
-			return nil, fmt.Errorf("duplicate door code found in %s: %s", filePath, door.Code)
-		}
-		door.Code = code
-		doorMap[code] = door
-	}
-
-	return doorMap, nil
-}
-
-// LoginItem defines a single step in the configurable login sequence.
-type LoginItem struct {
-	Command     string `json:"command"`                // Required: LASTCALLS, ONELINERS, USERSTATS, NMAILSCAN, DISPLAYFILE, RUNDOOR, FASTLOGIN
-	Data        string `json:"data,omitempty"`         // Optional: command-specific data (filename, script path, etc.)
-	ClearScreen bool   `json:"clear_screen,omitempty"` // Optional: clear screen before this item (default false)
-	PauseAfter  bool   `json:"pause_after,omitempty"`  // Optional: show pause prompt after this item (default false)
-	SecLevel    int    `json:"sec_level,omitempty"`    // Optional: minimum security level required (default 0 = everyone)
-}
-
-// LoadLoginSequence loads the login sequence configuration from login.json.
-// If the file is missing, returns a default sequence matching legacy behavior.
-func LoadLoginSequence(configPath string) ([]LoginItem, error) {
-	filePath := filepath.Join(configPath, "login.json")
-	slog.Info("loading login sequence", "path", filePath)
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Info("login.json not found, using default login sequence", "path", filePath)
-			return defaultLoginSequence(), nil
-		}
-		return nil, fmt.Errorf("failed to read login sequence file %s: %w", filePath, err)
-	}
-
-	var items []LoginItem
-	if err := json.Unmarshal(data, &items); err != nil {
-		return nil, fmt.Errorf("failed to parse login sequence JSON from %s: %w", filePath, err)
-	}
-
-	// Normalize command names to uppercase
-	for i := range items {
-		items[i].Command = strings.ToUpper(items[i].Command)
-	}
-
-	slog.Info("loaded login sequence", "count", len(items))
-	return items, nil
-}
-
-// defaultLoginSequence returns the built-in default login sequence matching legacy behavior.
-func defaultLoginSequence() []LoginItem {
-	return []LoginItem{
-		{Command: "LASTCALLS"},
-		{Command: "ONELINERS"},
-		{Command: "USERSTATS"},
-	}
-}
-
-// LoadOneLiners loads oneliner strings from a JSON file.
-func LoadOneLiners(dataPath string) ([]string, error) { // Changed configPath to dataPath for clarity
-	filePath := filepath.Join(dataPath, "oneliners.dat") // Assume loading .dat from data path
-	slog.Info("loading oneliners", "path", filePath)
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Warn("oneliners.dat not found, no oneliners loaded", "path", filePath)
-			return []string{}, nil // Return empty slice, not an error
-		}
-		slog.Error("failed to read oneliners file", "path", filePath, "error", err)
-		return nil, fmt.Errorf("failed to read oneliners file %s: %w", filePath, err)
-	}
-
-	// Assuming oneliners.dat is plain text, one per line
-	lines := strings.Split(strings.ReplaceAll(string(data), "\r\n", "\n"), "\n")
-	// Filter out potential empty lines
-	var oneLiners []string
-	for _, line := range lines {
-		if strings.TrimSpace(line) != "" {
-			oneLiners = append(oneLiners, line)
-		}
-	}
-
-	slog.Info("loaded oneliners", "count", len(oneLiners))
-	return oneLiners, nil
-}
-
-// ThemeConfig holds theme-related settings, loaded per menu set.
-// Colors are standard DOS color codes (0-255).
-type ThemeConfig struct {
-	YesNoHighlightColor int `json:"yesNoHighlightColor"`
-	YesNoRegularColor   int `json:"yesNoRegularColor"`
-	// Add other theme elements here as needed (e.g., default menu colors)
-}
-
-// LoadThemeConfig loads theme settings from theme.json within a specific menu set path.
-func LoadThemeConfig(menuSetPath string) (ThemeConfig, error) {
-	filePath := filepath.Join(menuSetPath, "theme.json")
-	slog.Info("loading theme configuration", "path", filePath)
-
-	// Default theme settings
-	defaultTheme := ThemeConfig{
-		YesNoHighlightColor: 112, // White on Black (inverse)
-		YesNoRegularColor:   15,  // Bright White on Black
-	}
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Warn("theme.json not found, using default theme settings", "path", filePath)
-			return defaultTheme, nil // Return defaults if file doesn't exist
-		}
-		slog.Error("failed to read theme file", "path", filePath, "error", err)
-		return defaultTheme, fmt.Errorf("failed to read theme file %s: %w", filePath, err)
-	}
-
-	// Initialize theme with defaults before unmarshalling
-	theme := defaultTheme
-	err = json.Unmarshal(data, &theme)
-	if err != nil {
-		slog.Error("failed to parse theme JSON, using default theme settings", "path", filePath, "error", err)
-		return defaultTheme, fmt.Errorf("failed to parse theme JSON from %s: %w", filePath, err)
-	}
-
-	slog.Info("loaded theme configuration", "path", filePath)
-	return theme, nil
-}
-
-// FTNLinkConfig defines an FTN link (uplink/downlink node).
-// Echo area routing is derived from message_areas.json (areas where Network matches),
-// not stored per-link. The Message Areas TUI is the canonical place to manage subscriptions.
-type FTNLinkConfig struct {
-	Address         string `json:"address"`                    // e.g., "21:1/100"
-	PacketPassword  string `json:"packet_password"`            // Packet password (formerly "password")
-	SessionPassword string `json:"session_password,omitempty"` // BinkP session/connection password
-	AreafixPassword string `json:"areafix_password,omitempty"` // Password for AreaFix netmail (subject line)
-	Name            string `json:"name"`                       // Human-readable name
-	Flavour         string `json:"flavour,omitempty"`          // Delivery flavour: Normal (default), Crash, Hold, Direct
-	Hostname        string `json:"hostname,omitempty"`         // Hub BinkP hostname; source of truth for the binkd.conf node line
-	Port            int    `json:"port,omitempty"`             // Hub BinkP port (default 24554 when Hostname is set)
-}
-
-// HostPort returns "hostname:port" for the link, defaulting the port to
-// 24554. Empty when no hostname is configured.
-func (c FTNLinkConfig) HostPort() string {
-	if c.Hostname == "" {
-		return ""
-	}
-	port := c.Port
-	if port <= 0 {
-		port = 24554
-	}
-	return fmt.Sprintf("%s:%d", c.Hostname, port)
-}
-
-// UnmarshalJSON supports backward compatibility: "password" is read into PacketPassword
-// when packet_password is absent (nil pointer = field omitted vs explicitly empty string).
-func (c *FTNLinkConfig) UnmarshalJSON(data []byte) error {
-	var r struct {
-		Address         string  `json:"address"`
-		PacketPassword  *string `json:"packet_password"`
-		SessionPassword string  `json:"session_password,omitempty"`
-		AreafixPassword string  `json:"areafix_password,omitempty"`
-		Name            string  `json:"name"`
-		Flavour         string  `json:"flavour,omitempty"`
-		Hostname        string  `json:"hostname,omitempty"`
-		Port            int     `json:"port,omitempty"`
-		LegacyPassword  string  `json:"password"`
-	}
-	if err := json.Unmarshal(data, &r); err != nil {
-		return err
-	}
-	c.Address = r.Address
-	c.SessionPassword = r.SessionPassword
-	c.AreafixPassword = r.AreafixPassword
-	c.Name = r.Name
-	c.Flavour = r.Flavour
-	c.Hostname = r.Hostname
-	c.Port = r.Port
-	if r.PacketPassword != nil {
-		c.PacketPassword = *r.PacketPassword
-	} else if r.LegacyPassword != "" {
-		c.PacketPassword = r.LegacyPassword
-	}
-	return nil
-}
-
-// FTNNetworkConfig holds settings for a single FTN network (e.g., FSXNet, FidoNet).
-// Netmail routing is derived from message_areas.json (areas where Network matches and AreaType == "netmail").
-type FTNNetworkConfig struct {
-	InternalTosserEnabled bool            `json:"internal_tosser_enabled"` // Enable internal tosser
-	OwnAddress            string          `json:"own_address"`             // e.g., "21:4/158.1"
-	PollSeconds           int             `json:"poll_interval_seconds"`   // 0 = manual only (v3mail toss/scan)
-	Tearline              string          `json:"tearline,omitempty"`      // Custom tearline text for echomail
-	Links                 []FTNLinkConfig `json:"links"`
-}
-
-// BinkdServerConfig controls the integrated binkd mailer daemon.
-// Zero-valued numeric/string fields are filled with defaults by LoadFTNConfig.
-type BinkdServerConfig struct {
-	Enabled    bool   `json:"enabled"`                 // Run binkd as a supervised child of the BBS
-	Port       int    `json:"port"`                    // binkp listen port (default 24554)
-	BinaryPath string `json:"binary_path"`             // Path to binkd binary, relative to BBS root (default "bin/binkd")
-	LogLevel   int    `json:"log_level"`               // binkd loglevel (default 4)
-	ExportSecs int    `json:"export_interval_seconds"` // Outbound scan/pack cadence (default 300)
-}
-
-// FTNConfig holds all FTN (FidoNet Technology Network) echomail settings.
-// Loaded from configs/ftn.json.
-type FTNConfig struct {
-	DupeDBPath        string                      `json:"dupe_db_path"`                  // e.g., "data/ftn/dupes.json"
-	InboundPath       string                      `json:"inbound_path"`                  // Where binkd deposits received bundles
-	SecureInboundPath string                      `json:"secure_inbound_path,omitempty"` // Authenticated inbound
-	OutboundPath      string                      `json:"outbound_path"`                 // Staging dir for outbound .PKT files
-	BinkdOutboundPath string                      `json:"binkd_outbound_path"`           // Binkd outbound dir for ZIP bundles
-	TempPath          string                      `json:"temp_path"`                     // Temp dir for processing
-	BadAreaTag        string                      `json:"bad_area_tag,omitempty"`        // Area for unroutable messages (e.g., "BAD")
-	DupeAreaTag       string                      `json:"dupe_area_tag,omitempty"`       // Area for duplicate messages (e.g., "DUPE")
-	Binkd             BinkdServerConfig           `json:"binkd"`                         // Integrated binkd mailer daemon
-	Networks          map[string]FTNNetworkConfig `json:"networks"`
-}
-
-// ServerConfig defines server-wide settings
-type ServerConfig struct {
-	BoardName           string `json:"boardName"`
-	SysOpName           string `json:"sysOpName"`
-	QWKID               string `json:"qwkID,omitempty"` // Explicit QWK packet ID; blank = derive from BoardName
-	BBSLocation         string `json:"bbsLocation,omitempty"`
-	Timezone            string `json:"timezone,omitempty"`
-	SysOpLevel          int    `json:"sysOpLevel"`
-	CoSysOpLevel        int    `json:"coSysOpLevel"`
-	WFCEnabled          bool   `json:"wfcEnabled"`     // Allow remote WFC sysop console (wfc-admin subsystem)
-	InvisibleLevel      int    `json:"invisibleLevel"` // Access level for invisible logon prompt; 0 = use coSysOpLevel
-	NewUserLevel        int    `json:"newUserLevel"`   // Access level assigned to new signups
-	RegularUserLevel    int    `json:"regularUserLevel"`
-	LogonLevel          int    `json:"logonLevel"`
-	AnonymousLevel      int    `json:"anonymousLevel"`
-	SSHPort             int    `json:"sshPort"`
-	SSHHost             string `json:"sshHost"`
-	SSHEnabled          bool   `json:"sshEnabled"`
-	TelnetPort          int    `json:"telnetPort"`
-	TelnetHost          string `json:"telnetHost"`
-	TelnetEnabled       bool   `json:"telnetEnabled"`
-	MaxNodes            int    `json:"maxNodes"`
-	MaxConnectionsPerIP int    `json:"maxConnectionsPerIP"`
-	IPBlocklistPath     string `json:"ipBlocklistPath"`
-	IPAllowlistPath     string `json:"ipAllowlistPath"`
-	MaxFailedLogins     int    `json:"maxFailedLogins"`
-	LockoutMinutes      int    `json:"lockoutMinutes"`
-	FileListingMode     string `json:"fileListingMode"`
-	LegacySSHAlgorithms bool   `json:"legacySSHAlgorithms"`
-	AllowNewUsers       bool   `json:"allowNewUsers"`
-
-	// Challenge Gate — optional pre-login bot challenge (botgate-style).
-	EnableChallengeGate          bool   `json:"enableChallengeGate"`          // master on/off
-	ChallengeGateFile            string `json:"challengeGateFile"`            // art file in menus ansi dir
-	ChallengeGateKey             string `json:"challengeGateKey"`             // "ESC" or a single character
-	ChallengeGateTimeoutSeconds  int    `json:"challengeGateTimeoutSeconds"`  // seconds to complete the challenge
-	ChallengeGateRequiredPresses int    `json:"challengeGateRequiredPresses"` // presses of the key to pass
-	ChallengeGateLiveCountdown   bool   `json:"challengeGateLiveCountdown"`   // live per-second countdown vs static
-
-	// Connection-rate limiter — temp-ban IPs that reconnect too rapidly.
-	EnableConnRateLimit        bool `json:"enableConnRateLimit"`
-	ConnRateLimitHits          int  `json:"connRateLimitHits"`          // attempts within window that trigger a ban
-	ConnRateLimitWindowSeconds int  `json:"connRateLimitWindowSeconds"` // sliding window size
-	ConnRateLimitBanMinutes    int  `json:"connRateLimitBanMinutes"`    // temp-ban duration
-
-	// Idle timeout (0 = disabled). Applied across the entire app; any input loop
-	// that calls ReadKeyWithTimeout uses this value.
-	SessionIdleTimeoutMinutes int `json:"sessionIdleTimeoutMinutes"`
-
-	// Transfer timeout in minutes for file transfers (ZModem, etc.). 0 = no timeout.
-	// When exceeded, the transfer process is killed and the session returns to the BBS.
-	TransferTimeoutMinutes int `json:"transferTimeoutMinutes"`
-
-	// DataDir is the runtime data directory (set by main at startup, not from JSON).
-	DataDir string `json:"-"`
-
-	// Number of days to retain soft-deleted user accounts before they are eligible
-	// for permanent purge. 0 = purge immediately; -1 = never purge automatically.
-	DeletedUserRetentionDays int `json:"deletedUserRetentionDays"`
-
-	// New User Voting (NUV) — community-based new user approval (V2 NUV system).
-	UseNUV      bool `json:"useNuv"`      // enable NUV system
-	AutoAddNUV  bool `json:"autoAddNuv"`  // automatically add new registrants to NUV queue
-	NUVUseLevel int  `json:"nuvUseLevel"` // minimum access level required to vote
-	NUVYesVotes int  `json:"nuvYesVotes"` // yes votes required to auto-validate
-	NUVNoVotes  int  `json:"nuvNoVotes"`  // no votes required to auto-delete
-	NUVValidate bool `json:"nuvValidate"` // auto-validate user when yes threshold reached
-	NUVKill     bool `json:"nuvKill"`     // auto-delete user when no threshold reached
-	NUVLevel    int  `json:"nuvLevel"`    // access level assigned on NUV auto-validation
-	NUVForm     int  `json:"nuvForm"`     // infoform number (1-5) to display during NUV voting; 0 = disabled
-
-	// DOS emulation — system-wide dosemu2 settings for DOS door games.
-	DosemuPath string `json:"dosemuPath,omitempty"` // Path to dosemu2 binary (default: /usr/libexec/dosemu2/dosemu2.bin)
-
-	// QWKAPI configures the optional QWK packet transport HTTP API (Phase 7).
-	// Disabled by default.
-	QWKAPI QWKAPIConfig `json:"qwkAPI"`
-
-	// Logging configuration (file location, level, caching, rolling). Shared by
-	// all binaries; each supplies its own default log filename in code.
-	Logging LoggingConfig `json:"logging"`
-}
-
-// Log rolling types persisted in LoggingConfig.Type.
-const (
-	LogTypeNone  = 0 // append to a single file indefinitely
-	LogTypeSize  = 1 // roll by size, keeping MaxFiles numbered backups
-	LogTypeDaily = 2 // roll daily, keeping MaxFiles days of dated files
-)
-
-// Default logging values, applied when the "logging" key is absent or a field
-// is left at its zero value.
-const (
-	DefaultLogDir       = "data/logs"
-	DefaultLogLevel     = "INFO"
-	DefaultLogMaxFiles  = 5
-	DefaultLogMaxSizeKB = 1024 // 1 MB
-)
-
-// LoggingConfig controls log file location, minimum level, write caching, and
-// rolling behavior. It is persisted under the "logging" key in config.json.
-type LoggingConfig struct {
-	Dir       string `json:"dir"`       // directory for log files (default "data/logs")
-	Level     string `json:"level"`     // DEBUG|INFO|WARN|ERROR minimum level (default INFO)
-	Cache     bool   `json:"cache"`     // buffer writes in an 8KB cache (default true)
-	Type      int    `json:"type"`      // 0=none 1=size 2=daily (default 0)
-	MaxFiles  int    `json:"maxFiles"`  // retained backups (type 1) / days (type 2)
-	MaxSizeKB int    `json:"maxSizeKB"` // rotate threshold in KB (type 1)
-}
-
-// Normalize fills empty strings with defaults and clamps numeric fields to safe
-// lower bounds so a config typo (e.g. MaxSizeKB=0) cannot cause immediate log
-// loss or rotate-on-every-write. The level string is validated separately by
-// the logging package at Init time.
-func (c *LoggingConfig) Normalize() {
-	if strings.TrimSpace(c.Dir) == "" {
-		c.Dir = DefaultLogDir
-	}
-	if strings.TrimSpace(c.Level) == "" {
-		c.Level = DefaultLogLevel
-	}
-	// MaxFiles bounds the number of retained backups (type 1) or days (type 2);
-	// it must be positive whenever rolling is enabled.
-	if c.Type != LogTypeNone && c.MaxFiles < 1 {
-		c.MaxFiles = DefaultLogMaxFiles
-	}
-	// MaxSizeKB only applies to size-based rolling; a non-positive threshold
-	// would rotate on every write.
-	if c.Type == LogTypeSize && c.MaxSizeKB < 1 {
-		c.MaxSizeKB = DefaultLogMaxSizeKB
-	}
-}
-
-// V3NetConfig holds V3Net networking configuration.
-type V3NetConfig struct {
-	Enabled      bool   `json:"enabled"`
-	KeystorePath string `json:"keystorePath"` // Path to ed25519 keypair file
-	DedupDBPath  string `json:"dedupDbPath"`  // Path to dedup SQLite database
-	RegistryURL  string `json:"registryUrl"`  // Central registry URL (optional)
-	// ConfigPath is the configs directory path. Set at runtime, not persisted.
-	ConfigPath string            `json:"-"`
-	Hub        V3NetHubConfig    `json:"hub,omitempty"`
-	Leaves     []V3NetLeafConfig `json:"leaves,omitempty"`
-}
-
-// QWKAPIConfig configures the optional QWK packet transport API (Phase 7).
-// Disabled by default. See docs/sysop/messages/qwk-api.md.
-type QWKAPIConfig struct {
-	Enabled       bool   `json:"enabled"`
-	Host          string `json:"host"`          // blank = all interfaces
-	Port          int    `json:"port"`          // default 8666
-	CertFile      string `json:"certFile"`      // blank = auto self-signed
-	KeyFile       string `json:"keyFile"`       // blank = auto self-signed
-	TokenTTLHours int    `json:"tokenTTLHours"` // default 24
-}
-
-// ListenAddr returns host:port for http.Server, defaulting the port to 8666.
-func (c *QWKAPIConfig) ListenAddr() string {
-	port := c.Port
-	if port == 0 {
-		port = 8666
-	}
-	return net.JoinHostPort(c.Host, strconv.Itoa(port))
-}
-
-// TokenTTL returns the bearer-token lifetime, defaulting to 24h.
-func (c *QWKAPIConfig) TokenTTL() time.Duration {
-	if c.TokenTTLHours <= 0 {
-		return 24 * time.Hour
-	}
-	return time.Duration(c.TokenTTLHours) * time.Hour
-}
-
-// V3NetHubConfig configures this node as a V3Net hub.
-type V3NetHubConfig struct {
-	Enabled      bool              `json:"enabled"`
-	Host         string            `json:"host"` // Listen host (blank = all interfaces)
-	Port         int               `json:"port"` // Listen port (default: 8765)
-	DataDir      string            `json:"dataDir"`
-	AutoApprove  bool              `json:"autoApprove"`
-	Networks     []V3NetHubNetwork `json:"networks,omitempty"`
-	InitialAreas []V3NetHubArea    `json:"initialAreas,omitempty"`
-}
-
-// ListenAddr returns the host:port string for net.Listen / http.Server.
-func (c *V3NetHubConfig) ListenAddr() string {
-	port := c.Port
-	if port == 0 {
-		port = 8765
-	}
-	return fmt.Sprintf("%s:%d", c.Host, port)
-}
-
-// V3NetHubNetwork defines a network hosted by this hub.
-type V3NetHubNetwork struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-}
-
-// V3NetHubArea is an area spec written by the setup wizard and consumed
-// once at hub startup to seed the initial NAL.
-type V3NetHubArea struct {
-	Tag         string `json:"tag"`
-	Name        string `json:"name"`
-	Description string `json:"description,omitempty"`
-}
-
-// V3NetLeafConfig configures a subscription to a V3Net network.
-type V3NetLeafConfig struct {
-	HubURL        string   `json:"hubUrl"`
-	Network       string   `json:"network"`
-	Boards        []string `json:"boards"`                  // Local message area tags to write received messages
-	PollInterval  string   `json:"pollInterval"`            // Duration string (e.g., "5m")
-	Origin        string   `json:"origin,omitempty"`        // Origin line text (e.g. "My Cool BBS - bbs.example.com")
-	AutoJoinAreas *bool    `json:"autoJoinAreas,omitempty"` // Flag created areas as newscan defaults (nil = true)
-}
-
-// AutoJoinEnabled reports whether message areas created for this leaf
-// should be flagged AutoJoin (newscan default). Unset means enabled, so
-// existing configs keep today's behavior.
-func (l V3NetLeafConfig) AutoJoinEnabled() bool {
-	return l.AutoJoinAreas == nil || *l.AutoJoinAreas
-}
-
-// EventConfig defines a scheduled event configuration
-type EventConfig struct {
-	ID                string            `json:"id"`
-	Name              string            `json:"name"`
-	Schedule          string            `json:"schedule"` // Cron syntax (empty for startup-only events)
-	Command           string            `json:"command"`
-	Args              []string          `json:"args"`
-	WorkingDirectory  string            `json:"working_directory"`
-	TimeoutSeconds    int               `json:"timeout_seconds"` // 0 = no timeout (for daemons)
-	Enabled           bool              `json:"enabled"`
-	RunAtStartup      bool              `json:"run_at_startup,omitempty"` // Launch immediately when BBS starts
-	EnvironmentVars   map[string]string `json:"environment_vars,omitempty"`
-	RunAfter          string            `json:"run_after,omitempty"`           // Event ID to run after
-	DelayAfterSeconds int               `json:"delay_after_seconds,omitempty"` // Delay after RunAfter completes
-}
-
-// EventsConfig is the root configuration for the event scheduler
-type EventsConfig struct {
-	// Enabled is deprecated and ignored: the scheduler always runs, and
-	// events are enabled/disabled individually. The field is kept so older
-	// binaries reading the same events.json still start their scheduler.
-	Enabled             bool          `json:"enabled"`
-	MaxConcurrentEvents int           `json:"max_concurrent_events"`
-	Events              []EventConfig `json:"events"`
-}
-
-// LoadServerConfig loads the server configuration from config.json
-func LoadServerConfig(configPath string) (ServerConfig, error) {
-	filePath := filepath.Join(configPath, "config.json")
-	slog.Info("loading server configuration", "path", filePath)
-
-	// Default config values
-	defaultConfig := ServerConfig{
-		BoardName:                    "ViSiON/3 BBS",
-		Timezone:                     "",
-		SysOpLevel:                   255,
-		CoSysOpLevel:                 250,
-		WFCEnabled:                   true,
-		NewUserLevel:                 1,
-		RegularUserLevel:             10,
-		LogonLevel:                   10,
-		AnonymousLevel:               5,
-		SSHPort:                      2222,
-		SSHHost:                      "0.0.0.0",
-		SSHEnabled:                   true,
-		TelnetPort:                   2323,
-		TelnetHost:                   "0.0.0.0",
-		TelnetEnabled:                false,
-		MaxNodes:                     10,
-		MaxConnectionsPerIP:          3,
-		MaxFailedLogins:              5,
-		LockoutMinutes:               30,
-		AllowNewUsers:                true,
-		EnableChallengeGate:          false,
-		ChallengeGateFile:            "BOTCHECK.ASC",
-		ChallengeGateKey:             "ESC",
-		ChallengeGateTimeoutSeconds:  20,
-		ChallengeGateRequiredPresses: 2,
-		ChallengeGateLiveCountdown:   true,
-		EnableConnRateLimit:          false,
-		ConnRateLimitHits:            20,
-		ConnRateLimitWindowSeconds:   10,
-		ConnRateLimitBanMinutes:      90,
-		SessionIdleTimeoutMinutes:    5,
-		TransferTimeoutMinutes:       10,
-		LegacySSHAlgorithms:          true,
-		DeletedUserRetentionDays:     30,
-		UseNUV:                       false,
-		AutoAddNUV:                   false,
-		NUVUseLevel:                  25,
-		NUVYesVotes:                  5,
-		NUVNoVotes:                   5,
-		NUVValidate:                  true,
-		NUVKill:                      false,
-		NUVLevel:                     25,
-		NUVForm:                      1,
-		Logging: LoggingConfig{
-			Dir:       DefaultLogDir,
-			Level:     DefaultLogLevel,
-			Cache:     true,
-			Type:      LogTypeNone,
-			MaxFiles:  DefaultLogMaxFiles,
-			MaxSizeKB: DefaultLogMaxSizeKB,
-		},
-	}
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Warn("config.json not found, using default settings", "path", filePath)
-			return defaultConfig, nil
-		}
-		return defaultConfig, fmt.Errorf("failed to read config file %s: %w", filePath, err)
-	}
-
-	// Initialize with defaults before unmarshalling
-	config := defaultConfig
-	err = json.Unmarshal(data, &config)
-	if err != nil {
-		slog.Error("failed to parse config JSON, using default settings", "path", filePath, "error", err)
-		return defaultConfig, fmt.Errorf("failed to parse config JSON from %s: %w", filePath, err)
-	}
-
-	config.SanitizeChallengeGate()
-	slog.Info("loaded server configuration", "path", filePath)
-	return config, nil
-}
-
-// SanitizeChallengeGate fills invalid/zero Challenge Gate values with safe
-// defaults, and also clamps the connection-rate-limit fields. Called after
-// load so a hand-edited config.json can't disable the gate through
-// out-of-range numbers or empty required strings. Note this does not
-// guarantee the connection-rate limiter stays enabled: a negative
-// ConnRateLimitHits clamps to 0, and 0 (or lower) hits is treated as
-// "disabled" by SetConnRateLimit.
-func (c *ServerConfig) SanitizeChallengeGate() {
-	if c.ChallengeGateFile == "" {
-		c.ChallengeGateFile = "BOTCHECK.ASC"
-	}
-	if c.ChallengeGateKey == "" {
-		c.ChallengeGateKey = "ESC"
-	}
-	if c.ChallengeGateTimeoutSeconds < 1 {
-		c.ChallengeGateTimeoutSeconds = 20
-	}
-	if c.ChallengeGateRequiredPresses < 1 {
-		c.ChallengeGateRequiredPresses = 2
-	}
-	if c.ConnRateLimitHits < 0 {
-		c.ConnRateLimitHits = 0
-	}
-	if c.ConnRateLimitWindowSeconds < 1 {
-		c.ConnRateLimitWindowSeconds = 10
-	}
-	if c.ConnRateLimitBanMinutes < 1 {
-		c.ConnRateLimitBanMinutes = 90
-	}
-}
-
-// applyBinkdDefaults fills zero-valued BinkdServerConfig fields with defaults.
-func applyBinkdDefaults(c *BinkdServerConfig) {
-	if c.Port == 0 {
-		c.Port = 24554
-	}
-	if c.BinaryPath == "" {
-		c.BinaryPath = "bin/binkd"
-	}
-	if c.LogLevel == 0 {
-		c.LogLevel = 4
-	}
-	if c.ExportSecs == 0 {
-		c.ExportSecs = 300
-	}
-}
-
-// LoadFTNConfig loads FTN network configuration from ftn.json.
-// Returns an empty config (no networks) if the file does not exist.
-func LoadFTNConfig(configPath string) (FTNConfig, error) {
-	filePath := filepath.Join(configPath, "ftn.json")
-	slog.Info("loading FTN configuration", "path", filePath)
-
-	defaultConfig := FTNConfig{
-		Networks: make(map[string]FTNNetworkConfig),
-	}
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Info("ftn.json not found, FTN disabled", "path", filePath)
-			applyBinkdDefaults(&defaultConfig.Binkd)
-			return defaultConfig, nil
-		}
-		return defaultConfig, fmt.Errorf("failed to read FTN config file %s: %w", filePath, err)
-	}
-
-	var config FTNConfig
-	err = json.Unmarshal(data, &config)
-	if err != nil {
-		slog.Error("failed to parse FTN config JSON", "path", filePath, "error", err)
-		return defaultConfig, fmt.Errorf("failed to parse FTN config JSON from %s: %w", filePath, err)
-	}
-
-	if config.Networks == nil {
-		config.Networks = make(map[string]FTNNetworkConfig)
-	}
-
-	enabledCount := 0
-	for name, net := range config.Networks {
-		if net.InternalTosserEnabled {
-			enabledCount++
-			slog.Info("ftn network internal tosser enabled", "network", name, "address", net.OwnAddress)
-		}
-	}
-	slog.Info("loaded FTN configuration", "networks", len(config.Networks), "tosserEnabled", enabledCount)
-
-	applyBinkdDefaults(&config.Binkd)
-	return config, nil
-}
-
-// ValidateFTNConfig checks that all required global path fields are set for any
-// network that has internal_tosser_enabled=true. Call this before starting the
-// tosser, not during editing, so the config editor can open an incomplete config.
-func ValidateFTNConfig(cfg FTNConfig) error {
-	tosserEnabled := false
-	for _, net := range cfg.Networks {
-		if net.InternalTosserEnabled {
-			tosserEnabled = true
-			break
-		}
-	}
-	if !tosserEnabled {
-		return nil
-	}
-	type requiredPath struct {
-		field string
-		value string
-	}
-	required := []requiredPath{
-		{"inbound_path", cfg.InboundPath},
-		{"outbound_path", cfg.OutboundPath},
-		{"binkd_outbound_path", cfg.BinkdOutboundPath},
-		{"temp_path", cfg.TempPath},
-	}
-	for _, r := range required {
-		if strings.TrimSpace(r.value) == "" {
-			return fmt.Errorf("ftn.json: %q is required when internal_tosser_enabled is true", r.field)
-		}
-	}
-	return nil
-}
-
-// ResolvePaths makes the FTN path fields absolute by joining relative paths
-// against root (the BBS root directory). Empty and absolute paths are unchanged.
-func (c *FTNConfig) ResolvePaths(root string) {
-	resolve := func(p string) string {
-		if p == "" || filepath.IsAbs(p) {
-			return p
-		}
-		return filepath.Join(root, p)
-	}
-	c.InboundPath = resolve(c.InboundPath)
-	c.SecureInboundPath = resolve(c.SecureInboundPath)
-	c.OutboundPath = resolve(c.OutboundPath)
-	c.BinkdOutboundPath = resolve(c.BinkdOutboundPath)
-	c.TempPath = resolve(c.TempPath)
-	if c.DupeDBPath != "" {
-		c.DupeDBPath = resolve(c.DupeDBPath)
-	}
-}
-
-// LoadV3NetConfig loads V3Net networking configuration from v3net.json.
-// Returns a disabled config if the file does not exist.
-func LoadV3NetConfig(configPath string) (V3NetConfig, error) {
-	filePath := filepath.Join(configPath, "v3net.json")
-	slog.Info("loading V3Net configuration", "path", filePath)
-
-	defaultConfig := V3NetConfig{}
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Info("v3net.json not found, V3Net disabled", "path", filePath)
-			return defaultConfig, nil
-		}
-		return defaultConfig, fmt.Errorf("failed to read V3Net config file %s: %w", filePath, err)
-	}
-
-	var cfg V3NetConfig
-	if err := json.Unmarshal(data, &cfg); err != nil {
-		slog.Error("failed to parse V3Net config JSON", "path", filePath, "error", err)
-		return defaultConfig, fmt.Errorf("failed to parse V3Net config JSON from %s: %w", filePath, err)
-	}
-
-	slog.Info("loaded V3Net configuration", "enabled", cfg.Enabled, "hub", cfg.Hub.Enabled, "leaves", len(cfg.Leaves))
-
-	return cfg, nil
-}
-
-// SaveV3NetConfig writes the V3NetConfig back to v3net.json in the given configPath directory.
-func SaveV3NetConfig(configPath string, cfg V3NetConfig) error {
-	filePath := filepath.Join(configPath, "v3net.json")
-	data, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal V3Net config: %w", err)
-	}
-	if err := os.WriteFile(filePath, data, 0644); err != nil {
-		return fmt.Errorf("failed to write V3Net config to %s: %w", filePath, err)
-	}
-	return nil
-}
-
-// SaveServerConfig writes the ServerConfig back to config.json in the given configPath directory.
-func SaveServerConfig(configPath string, cfg ServerConfig) error {
-	filePath := filepath.Join(configPath, "config.json")
-	data, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal server config: %w", err)
-	}
-	if err := os.WriteFile(filePath, data, 0644); err != nil {
-		return fmt.Errorf("failed to write config file %s: %w", filePath, err)
-	}
-	slog.Info("server configuration saved", "path", filePath)
-	return nil
-}
-
-// LoadEventsConfig loads the event scheduler configuration from events.json
-func LoadEventsConfig(configPath string) (EventsConfig, error) {
-	filePath := filepath.Join(configPath, "events.json")
-	slog.Info("loading event scheduler configuration", "path", filePath)
-
-	defaultConfig := EventsConfig{
-		Enabled:             true,
-		MaxConcurrentEvents: 3,
-		Events:              []EventConfig{},
-	}
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Info("events.json not found, no events to schedule", "path", filePath)
-			return defaultConfig, nil
-		}
-		return defaultConfig, fmt.Errorf("failed to read events config file %s: %w", filePath, err)
-	}
-
-	var config EventsConfig
-	err = json.Unmarshal(data, &config)
-	if err != nil {
-		slog.Error("failed to parse events config JSON", "path", filePath, "error", err)
-		return defaultConfig, fmt.Errorf("failed to parse events config JSON from %s: %w", filePath, err)
-	}
-
-	// Set default max concurrent events if not specified
-	if config.MaxConcurrentEvents <= 0 {
-		config.MaxConcurrentEvents = 3
-	}
-
-	enabledCount := 0
-	for _, event := range config.Events {
-		if event.Enabled {
-			enabledCount++
-		}
-	}
-
-	slog.Info("loaded event scheduler configuration", "events", len(config.Events), "enabled", enabledCount)
-
-	return config, nil
-}
 
 // LoadTimezone returns a *time.Location for the given timezone string.
 // It tries the value from config.json first, then the VISION3_TIMEZONE and TZ
